Fixing according to comments =)
